refactor: rename form component
