--- client/src/components/Schedule/AddEventModal/index.ts
@@ -1,1 +1,0 @@
-export { default as AddEventModal } from './AddEventModal';
