+++ client/src/components/Schedule/EventModalForm/index.ts
@@ -1,0 +1,1 @@
+export { default as EventModalForm } from './EventModalForm';
