+++ client/src/components/Schedule/index.ts
@@ -1,7 +1,7 @@
 export { ListView } from './ListView';
 export { TableView } from './TableView';
 export { CalendarView } from './CalendarView';
-export { AddEventModal } from './AddEventModal';
+export { EventModalForm } from './EventModalForm';
 export { EventDetails, TaskDetails } from './EventDetails';
 export { ScheduleSettings } from './ScheduleSettings';
 export { default as useScheduleSettings } from './useScheduleSettings';
