+++ client/src/pages/course/event.tsx
@@ -9,7 +9,7 @@ import { CourseService, CourseTaskDetails, CourseEvent } from '../../services/co
 import { CoursePageProps } from 'services/models';
 import { isCourseManager } from 'domain/user';
 import { TaskDetails, EventDetails } from 'components/Schedule/EventDetails';
-import { AddEventModal } from 'components/Schedule/AddEventModal';
+import { EventModalForm } from 'components/Schedule/EventModalForm';
 
 export function EventPage(props: CoursePageProps) {
   const { session, course } = props;
@@ -68,7 +68,7 @@ export function EventPage(props: CoursePageProps) {
         <EventDetails eventData={entityData as CourseEvent} alias={alias} onEdit={handleFullEdit} isAdmin={isAdmin} />
       )}
       {isModalOpen && (
-        <AddEventModal
+        <EventModalForm
           visible={isModalOpen}
           editableRecord={editableRecord as CourseEvent}
           handleCancel={closeModal}
